ViewEvent Move to Folder
