+++ src/app/calendar/page.js
@@ -22,8 +22,7 @@ import { RegionsContext } from '@/contexts/RegionsContext';
 import { useCalendarPage } from '@/hooks/useCalendarPage';
 import CalendarSubMenu from '@/components/UI/CalendarSubMenu';
 import CreateEventDetailModal from '@/components/Modals/CreateEvents/CreateEventDetailModal';
-
-import ViewEventDetailModal from '@/components/Modals/ViewEventDetailModal.js';
+import ViewEventDetailModal from '@/components/Modals/ViewEvents/ViewEventDetailModal.js';
 
 const CalendarPage = () => {
   <Head>
