+++ src/app/components/Modals/ViewEvents/ViewEventDetailsBasic.js
