+++ src/app/components/Modals/ViewEvents/ViewEventDetailsLocationOther.js
