+++ src/app/components/Modals/ViewEvents/ViewEventDetailsMore.js
