+++ src/app/components/Modals/ViewEvents/ViewEventDetailsRepeating.js
